README.md: Remove confusing assertion about `themes/wide16x9/`.
It's possible that the goal here was to mention `mods/wide16x9/`
instead, but the text still doesn't seem to describe the expected
behavior.  But I don't think it overrides `themes/default/` -- I think
it overrides any theme, not just the default.  So dropping it seems
simplest.

At any rate, if we want to guide the user about risks of using
`--mod=wide16x9` (or any mod), we should warn them in the
documentation for `--mod`, not in this side note about making
widescreen slides.